refactor: move resizing logic to utilClient
