+++ JWords/API/UtilClient.swift
@@ -7,10 +7,16 @@
 
 import ComposableArchitecture
 import XCTestDynamicOverlay
+#if os(iOS)
+import UIKit
+#elseif os(macOS)
+import AppKit
+#endif
 
 struct UtilClient {
     var filterOnlyFailUnits: ([StudyUnit]) -> [StudyUnit]
     var shuffleUnits: ([StudyUnit]) -> [StudyUnit]
+    var resizeImage: (InputImageType) -> InputImageType?
 }
 
 extension DependencyValues {
@@ -30,6 +36,40 @@ extension UtilClient: DependencyKey {
     },
     shuffleUnits: { units in
         units.shuffled()
+    },
+    resizeImage: { image in
+        // Calculate Size
+        let newWidth = Constants.Size.deviceWidth - 10
+        let newHeight = newWidth * (image.size.height / image.size.width)
+        let newSize = CGSize(width: newWidth, height: newHeight)
+        
+        // If image is small enough, return original one
+        if image.size.width < newWidth {
+            return image
+        }
+        
+        #if os(iOS)
+        UIGraphicsBeginImageContextWithOptions(newSize, false, 0.0)
+        image.draw(in: CGRect(origin: CGPoint.zero, size: newSize))
+        let resizedImage = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return resizedImage
+        #elseif os(macOS)
+         let newImage = NSImage(size: newSize)
+
+         newImage.lockFocus()
+
+         NSGraphicsContext.current?.imageInterpolation = .high
+
+         image.draw(in: NSRect(x: 0, y: 0, width: newWidth, height: newHeight),
+                    from: NSRect(x: 0, y: 0, width: image.size.width, height: image.size.height),
+                    operation: .sourceOver,
+                    fraction: 1.0)
+
+         newImage.unlockFocus()
+
+         return newImage
+         #endif
     }
   )
 }
@@ -37,7 +77,8 @@ extension UtilClient: DependencyKey {
 extension UtilClient: TestDependencyKey {
   static let previewValue = Self(
     filterOnlyFailUnits: { _ in .mock },
-    shuffleUnits: { _ in .mock }
+    shuffleUnits: { _ in .mock },
+    resizeImage: { _ in UIImage(named: "Sample Image") }
   )
 }
